+++ pkg/flow/logging/logger.go
@@ -126,9 +126,9 @@ func (l *Logger) Update(o Options) error {
 
 	// Print out the buffered logs since we determined the log format already
 	for _, bufferedLogChunk := range l.buffer {
-		if err := slogadapter.GoKit(l.handler).Log(bufferedLogChunk...); err != nil {
-			return err
-		}
+		// the buffered logs are currently only sent to the standard output
+		// because the components with the receivers are not running yet
+		slogadapter.GoKit(l.handler).Log(bufferedLogChunk...)
 	}
 	l.buffer = nil
 
@@ -164,6 +164,15 @@ type lokiWriter struct {
 
 func (fw *lokiWriter) Write(p []byte) (int, error) {
 	for _, receiver := range fw.f {
+		// We may have been given a nil value in rare circumstances due to
+		// misconfiguration or a component which generates exports after
+		// construction.
+		//
+		// Ignore nil values so we don't panic.
+		if receiver == nil {
+			continue
+		}
+
 		select {
 		case receiver.Chan() <- loki.Entry{
 			Labels: model.LabelSet{"component": "agent"},